Test Refactor - move TestColumn to top level class
[nostory]
