+++ server/src/test/java/org/cloudfoundry/identity/uaa/db/RevocableTokenTableTest.java
@@ -104,16 +104,4 @@ class RevocableTokenTableTest {
         return false;
     }
 
-    static class TestColumn {
-        final String name;
-        final String type;
-        final int size;
-
-        TestColumn(String name, String type, int size) {
-            this.name = name;
-            this.type = type;
-            this.size = size;
-        }
-    }
-
 }
